requirements.txt
crewai
sentence-transformers
faiss-cpu
watchfiles
pyyaml
mistune
